+++ .eslintrc.js
@@ -20,6 +20,11 @@ module.exports = {
     "no-console": 0,
     "indent": [2, 2, { "SwitchCase": 1, "MemberExpression": "off" }],
     "ember/no-jquery": 2,
+    // don't allow use of the global fetch
+    "no-restricted-globals": [2, {
+      name: "fetch",
+      message: "Use ember-fetch (import fetch from 'fetch') instead of the global."
+    }]
   },
   overrides: [
     // node files